+++ jovo-integrations/jovo-cms-googlesheets/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Current version might be higher than the latest changes displayed below because of updates of dependencies.
 
+### 2021-10-06 [3.5.2]
+* [#1059](https://github.com/jovotech/jovo-framework/pull/1059 ) :bug: Fix/google sheets after API deprecation ([@aswetlow](https://github.com/aswetlow))
+
 ## 3.0.11 (2020-04-06)
 
 * Updated Typescript to 3.8.x